# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Service General including Multi-DB, BPJS, and SatuSehat integrations.",
        "title": "GoPrint Service General API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/satusehat/reference/auth/token": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mendapatkan token aktif untuk API Satu Sehat Kemenkes (menggunakan cache internal)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Satu Sehat - Auth"
                ],
                "summary": "Get SatuSehat Access Token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/satusehat/reference/auth/token/refresh": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Memaksa request token baru dari API Satu Sehat Kemenkes (bypass cache)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Satu Sehat - Auth"
                ],
                "summary": "Refresh SatuSehat Access Token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/satusehat/reference/kfa/products": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil daftar produk KFA dengan kapabilitas paginasi",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Satu Sehat - KFA"
                ],
                "summary": "Daftar Produk KFA",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Nomor Halaman (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Jumlah Data (default: 10)",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Tipe Produk (farmasi/alkes)",
                        "name": "product_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Kata Kunci Pencarian",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Parameter waktu (from_)",
                        "name": "from_",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/satusehat/reference/kfa/products/{code}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mencari detail produk farmasi/alkes dari API Kamus Farmasi dan Alat Kesehatan (KFA) Satu Sehat",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Satu Sehat - KFA"
                ],
                "summary": "Cari Produk KFA berdasarkan Kode",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Kode KFA (contoh: 93000469)",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/satusehat/reference/location": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mencari data Location berdasarkan parameter",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Satu Sehat - Location"
                ],
                "summary": "Pencarian Lokasi (Satu Sehat)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nama Lokasi",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ID Organisasi Pemilik",
                        "name": "organization",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Identifier Lokasi",
                        "name": "identifier",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/satusehat/reference/location/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mencari data Location FHIR Satu Sehat berdasarkan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Satu Sehat - Location"
                ],
                "summary": "Cari Lokasi (Satu Sehat) berdasarkan ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Location ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/satusehat/reference/organization": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mencari data Organization berdasarkan Name atau PartOf",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Satu Sehat - Organization"
                ],
                "summary": "Pencarian Organisasi (Satu Sehat)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nama Organisasi",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ID Organisasi Induk",
                        "name": "partof",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Identifier Organisasi",
                        "name": "identifier",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/satusehat/reference/organization/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mencari data Organization FHIR Satu Sehat berdasarkan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Satu Sehat - Organization"
                ],
                "summary": "Cari Organisasi (Satu Sehat) berdasarkan ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                },
                "meta": {},
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}